+++ match-api/src/main/java/com/stocat/match/api/engine/MatchingEngine.java
@@ -1,6 +1,7 @@
 package com.stocat.match.api.engine;
 
 import com.stocat.match.domain.fill.Fill;
+import com.stocat.match.domain.fill.FillResult;
 import com.stocat.match.domain.order.Order;
 import com.stocat.match.domain.order.OrderType;
 import com.stocat.match.domain.orderbook.Orderbook;
